+++ packages/ethereum-indexer-browser/CHANGELOG.md
@@ -1,5 +1,12 @@
 # ethereum-indexer-browser
 
+## 0.6.21
+
+### Patch Changes
+
+- Updated dependencies
+  - ethereum-indexer@0.6.17
+
 ## 0.6.20
 
 ### Patch Changes